+++ aito/client/responses/__init__.py
@@ -2,5 +2,6 @@
 from .aito_response import BaseResponse, GetVersionResponse
 from .job_api_response import CreateJobResponse, GetJobStatusResponse
 from .query_api_response import BaseHit, ScoredHit, ProbabilityHit, RelateHit, HitsResponse, SearchResponse, \
-    PredictResponse, RecommendResponse, SimilarityResponse, MatchResponse, RelateResponse, EvaluateResponse
+    PredictResponse, RecommendResponse, SimilarityResponse, MatchResponse, RelateResponse, EvaluateResponse, \
+    EstimateResponse, AggregateResponse, ModifyResponse, BatchResponse
 from .schema_api_response import DatabaseSchemaResponse, TableSchemaResponse, ColumnSchemaResponse